argies/requirements.txt
fastapi==0.95.2
Jinja2==3.1.2
uvicorn==0.22.0
gunicorn==21.2.0
python-dateutil==2.8.2
python-long-weekends